Fix occasional seeks to NaN when running seekToStreamTime on live playlists
When handling video segment timing info in segment loader, the segment
was retrieved using the segment info object's playlist reference and
media index. However, if a live playlist refreshed during transmux, then
the playlist refrenced would be an old one, leaving the probed start/end
times on the updated segment, and the video timing info on the old
segment. Instead, use the segment info object's segment reference, which
is updated to point to the new segment on playlist refreshes.