fix(id3): ignore unsupported id3 frames (#437)
When we get an unsupported id3 frame, we don't parse its time. This ends
up coming back to us as NaN in the timing for the frame. Instead of
trying to create a cue for this frame, ignore it whenever the time isn't
a number or when it isn't finite.

Fixes videojs/video.js#5823