fix: end the updating state on receiving nothing back from the transmuxer (#610)
If the transmuxer never emits data but does signal that it is done with the current segment, we can finish handling this segment.
An example situation where this happens is when there is no audio or video content in a segment but there is timed metadata content. The transmuxer needs to be able to hold onto the metadata and wait for the next contentful segment. This change allows for that scenario.