fix: improve timestampOffset calculation to avoid gaps (#616)
Currently, the value for timestampOffset comes from segment.end, which is set from the DTS values returned by the tsprobe. This is not the correct value to use for the timestampOffset, and can result in missing frames or gaps between segments.

The TBA/LHLS branch (now #master) improves upon the current calculation of the timestampOffset by using the buffered end value instead of segment.end. However, this can still result in gaps in some cases if there is a difference between the pts time and dts time of the first frame that exceeds an internal browser threshold (in Chrome this threshold seems to be right around .08). The way to address that is to factor in the difference between the pts start time and dts start time of the first frame: timestampOffset = buffered.end - (ptsTime - dtsTime).