Simplify manifest modification functions by reducing abstraction
Previously, the same functions would be used for adding properties to
both master and media playlists. This simplified calls, but made for
more confusing sets of parameters and behaviors. It also led to some
harder to find backwards incompatibilities.

This simplifies the scope of each function to act on media playlists
regardless of whether they were part of a master or were sole media
playlists.