fix: race condition preventing qualityLevels from being populating (#707)
Quality levels is not being populated due to a race condition where the tech is not ready until after we have fired selectedintialmedia. You can get this to happen by in chrome by loading the tab in the background for 10s or by using a slow browser like ie 11, where it will happen in the main tab.

Fixes #677.