fix: workaround disappearing cues in non-segmented subtitles.
In DASH it's very common to provide the captions as one file rather than
having it be segmented VTT. It's possible to do the same in HLS, though,
it isn't common.

When we switch between tracks, we remove all the cues for the
duration of the video because the cues are no longer needed. Also, as we
re-download segments, the associated cues should be re-added.
Unfortunately, this the case of a single "segment", we don't get more
segments and the cues don't get re-added. Therefore, for non-live when
we have one segment, we shouldn't remove the cues.

This fixes the case both for turning the track on and off but also for
seeking.