fix: audio and video syncing for demuxed content with discontinuities
In order to properly sync audio and video for demuxed content when there
are discontinuities, one of the streams must follow the timing of the
other. For VHS, video was chosen (historically) to be the source of
truth for media timing, and audio was manipulated to fit the media
timing.

In order to properly fit audio to the video timeline, audio must know
the timing information of video before it is appended, and video must
wait for audio to cross a discontinuity before it crosses, so as not to
set the timestamp offset or other associated timing information for
audio before audio is done with appends along its current timeline.

As such, a TimelineChangeController object was added, in order to
facilitate discontinuity crossings where audio and video streams each
wait until the other is ready to cross before they make the move over to
a new timeline. This allows both to remain in sync, as they cross a
discontinuity together, and keep appropriate video timing information
for each timeline.