fix(segment-loader): resetEverything should remove through Infinity (#754) (#758)
In DASH it's very common to provide the captions as one file rather than
having it be segmented VTT. It's possible to do the same in HLS, though,
it isn't common.

When we switch between tracks, we remove all the cues for the
duration of the video because the cues are no longer needed. Also, as we
re-download segments, the associated cues should be re-added.
Unfortunately, this the case of a single "segment", we don't get more
segments and the cues don't get re-added. Therefore, for non-live when
we have one segment, we shouldn't remove the cues.

This fixes the case both for turning the track on and off but also for
seeking.

The fix here is to update resetEverything to call remove from `0` to `Infinity`. We then need to clamp the `end` value in the SegmentLoader due to issues in some older browsers.

Backported from #754.