chore(demo): add more sources, add keySystems and put DRMed sources in an optgroup (#828)
Add various sources. Include a keySystems textarea for working with DRM streams.

Support url structures with "empty" parts. Some sources like the unified-streaming source, includes a `/.mpd`, previously, with the `+` in the regex, this would've ignored that piece
and we wouldn't know this is an mpd. With the `*` we can properly detect the mpd file extension.