fix: Use partial xhr to detect sidx container and exclude non mp4 sidx (#824)
This pull request updates dash-playlist-loader to use a partial request before we would use sidx. This partial requests downloads as much data as needed to get the container type and ends. If the sidx segment is not mp4 it will be blacklisted. If it is mp4 sidx bytes will be passed along if they were downloaded. Otherwise we will do another request for sidx bytes.