feat: add experimental buffer based ABR (#886)
This adds a new option `experimentalBufferBasedABR` which turns it on. It uses a moving average playlist selector along with forward buffer checks that gate up and down switching. Right now it is turned off, and we are likely to fiddle with the implementation of this so it should not be relied on right now. We hope to eventually ship what this turns into as the new default ABR for VHS.

Co-authored-by: brandonocasey <brandonocasey@gmail.com>