fix: sustained bad network shouldn't pause
When playing back bipbop with experimentalBufferBasedABR turned on and
we switch to Slow 3G in chrome, we may time out.
This is because
- earlyabort is turned off
- our last attempt at downswitching occurs when our forward buffer is
too full

To mitigate this, when we get an early abort set up a timer for 10
seconds to see if earlyabort conditions are met again. If a progress
event happens in the meantime, the earlyabort timer will be cancelled.