Use a separate ProgramDateTime mapping per timeline
Previously, the ProgramDateTime would map to time 0 at the start of
playback, and that mapping would be used for the rest of playback. In
case of a discontinuity though, ProgramDateTime can jump (as in, can be
a larger difference in time than a single segment). This led to errors
while seeking. This change uses a mapping per timeline, allowing those
"jumps" in time on discontinuities.