fix: wait for endedtimeline event from transmuxer when reaching the end of a timeline (#1058)
Previously, when the end of a timeline was reached (i.e., the last
segment in a given timeline was appended), `endTimeline` would be
called, but the loader would not wait for the event to finish. This
lead to issues where segment information came back after a segment is
considered done with processing, and segment information could be
dropped. This fix waits for the `endedtimeline` message from the
transmuxer before it considers the transmux complete.