fix: use a separate ProgramDateTime mapping to player time per timeline (#1063)
Previously, the ProgramDateTime would map to time 0 at the start of playback, and that mapping would be used for the rest of playback.

However, in the case of a discontinuity, ProgramDateTime can jump (as in, can be a larger difference in time than a single segment duration, e.g., if the encoder went down for a period of time). Because the ProgramDateTime to player time mapping never changed, this led to seeking errors.

This fix uses a mapping of ProgramDateTime to player time per timeline, allowing those "jumps" in time on discontinuities.