fix: live dash segment changes should be considered a playlist update (#1065)
Currently dash content goes off of the number attribute for segments to determine the mediaSequence. This is problematic as the first segment is almost always listed as 1. So in essence when we request the playlist again the media sequence will be the same, and likely the amount of segments will be the same, so we will determine that this playlist update didn't update anything at all.