fix: don't clear DASH minimum update period timeout on pause of a media loader
Previously, for DASH, pausing a media loader would end up pausing the
main loader as well (via removal of the main loader's minimum update
period). For live streams which required manifest refreshes, this meant
that on media changes, the main loader would be paused and sometimes
never resumed.

This change ensures that when a media loader is paused it won't remove
the main loader's minimum update period timeout.

In addition, a change was made so that loading a main DASH playlist
loader recreates a minimum update period timeout if the main DASH loader
was previously paused (and the timeout cleared).