feat: allow clients to limit the number of times a playlist attempts to reload following an error (#1098)
Introduce a new option `maxPlaylistRetries`, which defaults to Infinity to match existing behavior. If set, this will cause a playlist to be excluded indefinitely when the threshold is crossed. When a playlist is requested successfully, the count is reset.

If a stream has multiple playlists, all playlists will need to cross the max retries threshold for playback to completely fail.

Co-authored-by: Evan farina <efarina@linkedin.com>
Co-authored-by: Gary Katsevman <git@gkatsev.com>
Co-authored-by: brandonocasey <brandonocasey@gmail.com>