Support HLS smoothQualityChange setting on MSE
The current "fast quality switching" policy which clears the buffers and replaces then as fast as possible with others with the new quality is good for fast quality switching on regular desktop systems, which have a software decoder and plenty of CPU.

However, on set-top-boxes using embedded browsers like [WPE WebKit](https://github.com/WebPlatformForEmbedded/WPEWebKit), this causes an internal "flush and reenqueue" on the playback pipeline, which is emptied and has the buffers with new quality enqueued and quickly decoded until the old playback position is reached again. This causes a big stress on the decoder to catch up, and it's even worse on devices with hardware decoders and custom audio-video synchronization mechanisms.

For this reason, those systems would benefit a lot on having the smoothQualityChange setting honored when using Media Source Extensions.

In order to use this feature, the user should configure the player like this:

  var player = videojs('videoelementid', {
    html5: {
      hls: {
        overrideNative: true,
        smoothQualityChange: true
      }
    }
  });