test: change source for live DASH playback test to fix test failures
The current live DASH playback test source uses a source where segments
are referenced by a SegmentTemplate and duration, but the server
doesn't have time syncing and appears to be about two minutes behind
local time. This leads to requests for segments beyond expected and
occasional failures.

Until that issue is fixed in VHS, switch to a live source that uses
SegmentTemplate, to be a more reliable basic live DASH playback test.