feat: Add experimental support for ManagedMediaSource (#1453)
Adds basic support for ManagedMediaSource. Must be enabled with the `useManagedMediaSource` VHS option.

Does not implement an alternate AirPlay source - this requires a more significant change, to add two source els. This means remote playback has to be disabled on the video el when using MMS.

Event listeners for advanced control are not yet implemented - `startstreaming`, `endstreaming`, `qualitychange`